Clean up `GolWorld`
* Add some documentation to `GolWorld.hpp`.
* Extract `wrapAround` to a new file under `src/util/`.
* Reorder implementations in `GolWorld.cpp`.
* Miscellaneous small touches.
