+++ src/scenes/gol/golWorld.cpp
@@ -4,21 +4,12 @@
 #include <cassert>
 #include <iostream>
 
-// TODO organize/reorder
-
-// TODO move this somewhere else?
-namespace mossman {
-	int wrapAround(int i, int max) {
-		auto mod = i % max;
-		return mod < 0 ? max + mod : mod;
-	}
-}
-
 GolWorld::GolWorld(int columns, int rows) 
 		: nCols(columns), nRows(rows) {
 	randomizeState();
 
 	// Build default neighbourhood
+	// TODO extract somewhere? need to consider neighbourhood implementations more
 	neighbourhood.reserve(8);
 	for(int i = -1; i <= 1; i++) {
 		for(int j = -1; j <= 1; j++) {
@@ -41,33 +32,14 @@ void GolWorld::step() {
 	this->cells = buffer;
 }
 
-int GolWorld::countAliveNeighbours(int x, int y) const {
-	int sum = 0;
-	int col, row;
-	for(auto neighbour : neighbourhood) {
-		col = wrapColumn(neighbour.first + x);
-		row = wrapRow(neighbour.second + y);
-		if(isAlive(col, row)) sum++;
-	}
-	return sum;
-}
-
-CellContainer GolWorld::buildCellContainer() const {
-	return gol::buildCellContainer(nCols, nRows);
+void GolWorld::resize(size_t columns, size_t rows) {
+	nCols = columns;
+	nRows = rows;
+	gol::resizeCellContainer(cells, nCols, nRows);
 }
 
 void GolWorld::randomizeState() {
 	this->cells = buildRandomizedCellContainer();
-}
-
-CellContainer GolWorld::buildRandomizedCellContainer() const {
-	auto buffer = buildCellContainer();
-	for(int x = 0; x < nCols; x++) {
-		for(int y = 0; y < nRows; y++) {
-			buffer[x][y] = (rand() % 2 == 0) ? true : false;
-		}
-	}
-	return buffer;
 }
 
 void GolWorld::setRules(RuleSet rules) {
@@ -84,4 +56,29 @@ void GolWorld::insertShape(const CellContainer& shape, int x, int y) {
 			cells[x+i][y+j] = shape[i][j];
 		}
 	}
+}
+
+int GolWorld::countAliveNeighbours(int x, int y) const {
+	int sum = 0;
+	int col, row;
+	for(auto neighbour : neighbourhood) {
+		col = wrapColumn(neighbour.first + x);
+		row = wrapRow(neighbour.second + y);
+		if(isAlive(col, row)) sum++;
+	}
+	return sum;
+}
+
+CellContainer GolWorld::buildCellContainer() const {
+	return gol::buildCellContainer(nCols, nRows);
+}
+
+CellContainer GolWorld::buildRandomizedCellContainer() const {
+	auto buffer = buildCellContainer();
+	for(int x = 0; x < nCols; x++) {
+		for(int y = 0; y < nRows; y++) {
+			buffer[x][y] = (rand() % 2 == 0) ? true : false;
+		}
+	}
+	return buffer;
 }
